postgresql-python/requirements.txt
argon2-cffi==20.1.0
asyncpg==0.21.0
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
cffi==1.14.2
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
import==0.0.1
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.6
psycopg2==2.8.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
PyGreSQL==5.2
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
Send2Trash==1.5.0
six==1.15.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
